Arreglos varios . - Arreglo de que un número de perforación pueda ser decimal - Arreglo calendario input, ahora se abre con un solo click. - Arreglo de los nombres del menú.

+++ app/src/main/java/com/example/proyecto_german/Model/PerforacionModel.kt
@@ -16,7 +16,7 @@ data class PerforacionModel(
     val proyecto:String,
     val localizacion:String,
     val fecha: Date,
-    val numeroPerforacion: Int,
+    val numeroPerforacion: Double,
     val profundidadMetros: Double,
     val coordenadaE:Double,
     val coordenadaN: Double,
